Second-order problems

This section includes direct integration methods for linear dynamic equations of the form

\[ Mx''(t) + Cx'(t) + Kx(t) = F(t)\]

In the context of structural dynamics problems, $M$ is the mass matrix, $C$ is the viscous damping matrix, $K$ is the stiffness matrix, $F$ is the vector of applied forces, and $x(t)$, $x'(t)$ and $x''(t)$ are the displacement, velocity and acceleration vectors, respectively.

The following algorithms are available:

Central difference method.

Houbolt's method.

Newmark's method.

Bathe integration method with equal-size substeps ($\gamma = 0.5$).

The theoretical description of such methods can be found in Chapter 9, [BATHE]; see also the references appearing in each docstring.

Central difference

StructuralDynamicsODESolvers.CentralDifference — Type

struct CentralDifference{N} <: StructuralDynamicsODESolvers.AbstractSolver

Central difference scheme with given step-size.

Fields

Δt – step-size

Houbolt

StructuralDynamicsODESolvers.Houbolt — Type

struct Houbolt{N} <: StructuralDynamicsODESolvers.AbstractSolver

Houbolt's integration scheme with given step-size.

Fields

Δt – step-size

References

See [HOU50].

Newmark

StructuralDynamicsODESolvers.Linear — Function

Linear(Δt::N)

Linear integration scheme. Special case of Newmark with $δ=1/2$ and $α=1/6$.

StructuralDynamicsODESolvers.Newmark — Type

struct Newmark{N} <: StructuralDynamicsODESolvers.AbstractSolver

Newmark's integration scheme with given step-size and parameters α and δ.

Fields

Δt – step-size

α – parameter α of the method

δ – parameter δ of the method

References

See [NEW59].

StructuralDynamicsODESolvers.Trapezoidal — Function

Trapezoidal(Δt::N)

Trapezoidal integration scheme. Special case of Newmark with $δ=1/2$ and $α=1/4$.

Bathe

StructuralDynamicsODESolvers.Bathe — Type

struct Bathe{N} <: StructuralDynamicsODESolvers.AbstractSolver

Bathe's integration scheme with sub-steps of equal size.

Fields

Δt – step-size

α – parameter α of the method

δ – parameter δ of the method

References

See [BAT07].